refactor: change axios to api
