+++ apps/app/src/renderer/requests/auth/confirm-code.ts
@@ -1,5 +1,5 @@
-import { api } from '@renderer/lib/axios'
 import type { ConfirmCodeRequest, ConfirmCodeResponse } from '@repo/api-types/authentication.dto'
+import { api } from '~/src/renderer/lib/api'
 
 export const confirmCode = async ({
 	email,
