+++ apps/app/src/renderer/requests/auth/get-code.ts
@@ -1,5 +1,5 @@
-import { api } from '@renderer/lib/axios'
 import type { GetCodeRequest, GetCodeResponse } from '@repo/api-types/authentication.dto'
+import { api } from '~/src/renderer/lib/api'
 
 export const getCode = async ({ email }: GetCodeRequest): Promise<GetCodeResponse> =>
 	api.post('/auth/get-code', { email })
