+++ apps/app/src/renderer/requests/payments/get-all-payments.ts
@@ -1,5 +1,5 @@
 import type { Payment } from '@repo/api-types/payment.dto'
-import { api } from '../../lib/axios'
+import { api } from '../../lib/api'
 
 export const getAllPayments = async (userId: string): Promise<Payment[]> =>
 	api.get(`/payment/all/${userId}`)
