+++ apps/app/src/renderer/requests/payments/get-payment.ts
@@ -1,5 +1,5 @@
 import type { Payment } from '@repo/api-types/payment.dto'
-import { api } from '../../lib/axios'
+import { api } from '../../lib/api'
 
 export const getPayment = async (paymentId: string): Promise<Payment> =>
 	await api.get(`/payment/pix/${paymentId}`)
