+++ apps/app/src/renderer/requests/user/get-user.ts
@@ -1,4 +1,4 @@
-import { api } from '@renderer/lib/axios'
 import type { User } from '@repo/api-types/user.schema'
+import { api } from '~/src/renderer/lib/api'
 
 export const getUser = async (userId: string): Promise<User> => await api.get(`/user/${userId}`)
